moved login handler to separate file
